styling primary, secondary text, and primary btn
