+++ client/package.json
@@ -37,5 +37,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  }, 
+  "devDependencies": {
+      "sass": "^1.49.9"
   }
 }
