+++ client/src/App.js
@@ -1,25 +1,25 @@
 import React from 'react';
 import { BrowserRouter as Router, Routes, Route, Link } from 'react-router-dom';
-import './App.css';
+import './App.scss';
 
-import { Signup } from './Signup';
-import { Login } from './Login';
+import { Signup } from './components/Signup';
+import { Login } from './components/Login';
 
 function App() {
   return (
-    <Router>
-      <div className="App">
-        <nav>
+    <div className="App">
+      <Router>
+        {/* <nav>
           <Link to="/signup">Sign Up</Link>
           <Link to="/">Login</Link>
-        </nav>
+        </nav> */}
 
         <Routes>
           <Route exact path="/" Component={Login} />
           <Route exact path="/signup" Component={Signup} />
         </Routes>
-      </div>
-    </Router>
+      </Router>
+    </div>
   );
 }
