+++ client/src/App.scss
@@ -4,6 +4,8 @@
     align-items: center;
     justify-content: center;
     background-color: white;
+    height: 100vh;
+    padding: 40px
 }
 
 .signup-container {
@@ -46,7 +48,6 @@
 
 .signup-header {
     color: #000;
-    font-family: Inter;
     font-size: 40px;
     font-weight: 400;
     align-items: center;
@@ -57,7 +58,6 @@ label {
     text-align: left;
     padding: 8px 0;
     color: #000;
-    font-family: Inter;
     font-size: 24px;
     font-weight: 400;
 }
@@ -72,29 +72,14 @@ input {
 
 input::placeholder {
     color: #000;
-    font-family: Inter;
     font-size: 12px;
     font-style: italic;
     font-weight: 400;
     opacity: 0.5;
 }
 
-button {
-    width: 148px;
-    height: 54px;
-    border-radius: 10px;
-    background: #F3B144;
-    color: #161C47;
-    text-align: center;
-    font-family: Inter;
-    font-size: 24px;
-    font-weight: 700;
-    cursor: pointer;
-}
-
 .switch-page {
     color: #000;
-    font-family: Inter;
     font-size: 24px;
     font-weight: 400;
 }
