+++ client/src/styles/app_variables.scss
@@ -1,0 +1,43 @@
+$primary: #F3B144;
+$secondary: #161C47;
+
+.btn-primary {
+    text-align: center;
+    border-radius: 10px !important;
+    
+    color: $secondary;
+    background-color: $primary;
+    border-color: transparent
+}
+
+.btn-primary:hover {
+    color: $secondary;
+    background-color: darken($primary, 7.5%);
+    border-color: transparent
+}
+
+.btn-primary:focus,
+.btn-primary.focus {
+    box-shadow: 0 0 0 .2rem rgba(mix($primary, $primary, 15%), .5)
+}
+
+.btn-primary.disabled,
+.btn-primary:disabled {
+    color: $secondary;
+    background-color: $primary;
+    border-color: $primary
+}
+
+.btn-primary:not(:disabled):not(.disabled):active,
+.btn-primary:not(:disabled):not(.disabled).active,
+.show>.btn-primary.dropdown-toggle {
+    color: $secondary;
+    background-color: darken($primary, 10%);
+    border-color: transparent
+}
+
+.btn-primary:not(:disabled):not(.disabled):active:focus,
+.btn-primary:not(:disabled):not(.disabled).active:focus,
+.show>.btn-primary.dropdown-toggle:focus {
+    box-shadow: 0 0 0 .2rem rgba(mix($primary, $primary, 15%), .5)
+}
