+++ client/src/styles/overrides.scss
@@ -1,0 +1,5 @@
+$theme-colors: (
+  "primary": #F3B144,
+  "secondary": #161C47 
+);
+$font-family-base: Inter;
